requirements.txt
streamlit==1.52.2
pandas==2.3.3
plotly==5.24.1
google-generativeai==0.8.3
pdfplumber==0.11.4
Pillow==12.0.0
passlib==1.7.4